% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstR-package.R
\docType{package}
\name{dstR-package}
\alias{dstR}
\alias{dstR-package}
\title{dstR: Access data From Statistics Denmark}
\description{
dstR provides an R interface to all openly available data from Statistics Denmark
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/elben10/dstR/issues}
}

}
\author{
\strong{Maintainer}: Jakob Jul Elben \email{elbenjakobjul@gmail.com}

}
