% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr_network.R
\name{dstr_network}
\alias{dstr_network}
\title{Create a dependency network object.}
\usage{
dstr_network(githublink = NULL, pkg = NULL, includebasepkgs = F,
  recursive = T)
}
\arguments{
\item{githublink}{Link of type character to a github repository of a R package.}

\item{pkg}{Character input of a packagename you want to plot the dependencies for.
Multiple input packages can be given in vector format.}

\item{includebasepkgs}{If TRUE all dependencies to base R packages are included
in the output. Take this option with caution. The output easily gets very chaotic.}

\item{recursive}{If TRUE the network graph will show dependencies of dependencies.}
}
\value{
Output will be an igraph network object which can be manipulated by standard
igraph rules.
}
\description{
\code{dstr_network()} returns an igraph network object, which can be used for
creating individual dependency graphs.
}
\examples{

network_object <- dstr_network(githublink = "falo0/dstr", pkg = "ggplot2")
#plot.igraph(network_object, edge.arrow.size = .1, edge.color="darkgrey",vertex.size = 10,
#           vertex.shape = "circle",vertex.frame.color = "white", vertex.label.font= 1,
#            vertex.label.color = "black", vertex.color = "white",edge.width = 1.5,
#            layout = layout_with_fr)

}
