% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstr_summary.R
\name{dstr_summary}
\alias{dstr_summary}
\title{Dependency Summary}
\usage{
dstr_summary(githublink = NULL, pkg = NULL, includebasepkgs = F)
}
\arguments{
\item{githublink}{A link to a github repository of an R package.}

\item{pkg}{A list of packages from which we want to know the further
dependencies. This list will be added to the first level dependencies
of a given package on github it githublink is set.}

\item{includebasepkgs}{Whether to include base packages in the analysis.}
}
\description{
This function gives a summary of the dependency structure of a given package
and points out opportunities to eliminate depdendenceis completely.
}
\details{
The default assumption is that there is an R package in the current working
directory and that the dependencies to be analyzed are given in the DESCRIPTION
file. Use the parameters ‘githublink’ and/or 'pkg' to alter the package/s
to be analyzed.
}
