% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstrsummary.R
\name{dstrsummary}
\alias{dstrsummary}
\title{Dependency Summary}
\usage{
dstrsummary(githublink = NULL, pkg = NULL)
}
\arguments{
\item{githublink}{A link to a github repository of an R package}

\item{pkg}{A list of packages from which we want to know the further
dependencies. This list will be added to the first level dependencies
of a given package on github it githublink is set.}
}
\description{
This function gives a summary of the dependency structure of a given package
and points out opportunities to eliminate depdendenceis completely.
}
