% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstlvldep.R
\name{firstlvldep}
\alias{firstlvldep}
\title{First Level Dependencies}
\usage{
firstlvldep(githublink = NULL, localdescdir = NULL)
}
\arguments{
\item{githublink}{A link to the package repository or directly to the description file}

\item{localdescdir}{A directory to a local DESCRIPTION file}
}
\description{
This function returns all first level packag dependencies from a package on github.
}
\examples{
firstlvldep("https://github.com/tidyverse/ggplot2")
}
