% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_pkgname_rawlink.R
\name{helper_pkgname_rawlink}
\alias{helper_pkgname_rawlink}
\title{Link To Raw DESCRIPTION, Package Name}
\usage{
helper_pkgname_rawlink(githublink)
}
\arguments{
\item{githublink}{A link to the package repository or directly to the description file}
}
\description{
This function inputs various ways to link to a package or directly the description file on github
It outputs the name of the package and a link to the raw DESCRIPTION file.
}
