% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nthlvldep.R
\name{nthlvldep}
\alias{nthlvldep}
\title{Get All Depper Level Dependencies}
\usage{
nthlvldep(pkg, outtype = "edge_list")
}
\arguments{
\item{pkg}{A list of packages from which we want to know the further
dependencies}

\item{outtype}{Possible output types:
edge_list: An edge list, e.g. to be used for network plots
allPkgs: An overview of all packages that are eventually loaded. No further
structure visible
list: More detailed than allPKs, it's a list tha containns all packages per
first level dependency
unique_list: like "list", just excluding all packages, that are eventually
also loaded by another package in firstlvldep. This way you can see which
dependencies will be removed completely if you remove a certain first level
dependency (a package that you import).
tree: Detailed information about which package depends on which, represented
in a data frame that is showing a tree structure.}
}
\description{
This function inputs a list of packages, regarded to as fist level
dependencies. It outputs the dependencies, dependencies of dependencies, ...
of these first level dependencies. There are various options for the type of
the output.
}
