% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges2tree.R
\name{edges2tree}
\alias{edges2tree}
\title{Convert Edge List To Data Frame With Tree Structure}
\usage{
edges2tree(edges, lvl1deps)
}
\arguments{
\item{edges}{An edge list as a data frame, first column: starting points, second columns: end points}

\item{lvl1deps}{The packages that should be at the first level (at the roots) of the tree structure}
}
\description{
This helper function inputs an edge list and converts it into a data frame
where the whole tree structure can be seen.
}
\examples{
edgelist <- data.frame(start = c("shiny", "shiny", "htmltools"),
   end = c("jsonlite", "htmltools", "digest"))
lvl1deps <- "shiny"
edges2tree(edgelist, lvl1deps)
}
