% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdstr.R
\name{plotdstr}
\alias{plotdstr}
\title{Plot Dependencs Structure}
\usage{
plotdstr(githublink = NULL, pkg = NULL, includebasepkgs = F,
  recursive = T)
}
\arguments{
\item{githublink}{A link to a github repository of an R package}

\item{pkg}{A list of packages from which we want to know the further
dependencies. This list will be added to the first level dependencies
of a given package on github it githublink is set}

\item{includebasepkgs}{Whether to include base packages in the analysis or not}

\item{recursive}{Whether you want to look deeper than the second level of dependencies,
e.g. get all dependencies of dependencies of dependencies ...}
}
\description{
This function plots the dependency structur of a package
}
